Update docker build steps to allow for caching of dependencies (#191)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description
<!--- Describe your changes in detail -->
Update docker build to allow for caching of installed dependencies.
Currently rebuilding CDASim takes excessively long because the docker
layer that pulls sumo and CARLA is the same one that builds mosiac.
To improve on this I suggest we use scripts like
**install_dependencies.sh** -> **build.sh** -> **run.sh** so that all
dependencies can be installed in a single build layer and cached and
changes to source code should not trigger dependencies to be downloaded
again and rebuilt. This PR only includes the addition of an
"install_dependencies.sh" and a "build.sh". There already exists a run
script called mosaic.sh. There may be some reorganization necessary to
allow the added scripts directory to coexist with the docker directory.
Would suggest further improvements for moving install scripts out of
docker directory and uncoupling them from docker deployment. In the
future, the docker director should only including things in this
directory that are specific to docker like tagging scripts.
## Related GitHub Issue

<!--- This project only accepts pull requests related to open GitHub
issues or Jira Keys -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please DO NOT name partially fixed issues, instead open an issue
specific to this fix -->
<!--- Please link to the issue here: -->

## Related Jira Key
[CDAR-719](https://usdot-carma.atlassian.net/browse/CDAR-719)
<!-- e.g. CAR-123 -->

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
Quicker builds
## How Has This Been Tested?
CI
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Defect fix (non-breaking change that fixes an issue)
- [x] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.


[CDAR-719]:
https://usdot-carma.atlassian.net/browse/CDAR-719?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ